# 3.spec-kit-openspec/java/backend/src/main/java/com/cms/presentation/dto/CreateCategoryRequest.java
package com.cms.presentation.dto;

public class CreateCategoryRequest {
    private String name;
    private String description;
    private Long parentId;
    private Integer displayOrder;

    // Getters and Setters
    public String getName() { return name; }
    public void setName(String name) { this.name = name; }
    public String getDescription() { return description; }
    public void setDescription(String description) { this.description = description; }
    public Long getParentId() { return parentId; }
    public void setParentId(Long parentId) { this.parentId = parentId; }
    public Integer getDisplayOrder() { return displayOrder; }
    public void setDisplayOrder(Integer displayOrder) { this.displayOrder = displayOrder; }
}